docker/hummingbot-mm/requirements.txt
# Deluthium Hummingbot Market Maker Dependencies

# WebSocket client
websockets>=12.0

# Ethereum signing
eth-account>=0.11.0
eth-typing>=4.0.0

# Configuration
pyyaml>=6.0

# HTTP client (for price feeds)
aiohttp>=3.9.0

# Utilities
python-dotenv>=1.0.0
